+++ src/test/scala/org/kibanaLoadTest/simulation/local/AtOnceJourney.scala
@@ -5,51 +5,32 @@ import io.gatling.core.Predef._
 import io.gatling.core.structure.ScenarioBuilder
 import org.kibanaLoadTest.scenario.{Canvas, Dashboard, Discover}
 
-import java.util.concurrent.TimeUnit
-import scala.concurrent.duration.FiniteDuration
-
 class AtOnceJourney extends BaseSimulation {
   def scenarioName(module: String): String = {
     s"Cloud  atOnce ${module} ${appConfig.buildVersion}"
   }
 
-  val simulationTimeout =
-    FiniteDuration(
-      10,
-      TimeUnit.MINUTES
-    )
-
-  val loginPause =
-    FiniteDuration(
-      2,
-      TimeUnit.SECONDS
-    )
-
-  val scnPause =
-    FiniteDuration(
-      20,
-      TimeUnit.SECONDS
-    )
+  props.maxUsers = 400
 
   val scnDiscover: ScenarioBuilder = scenario(scenarioName("discover"))
-    .exec(loginStep.pause(loginPause))
+    .exec(loginStep.pause(props.loginPause))
     .exec(Discover.doQuery(appConfig.baseUrl, defaultHeaders))
 
   val scnDashboard: ScenarioBuilder = scenario(scenarioName("dashboard"))
-    .exec(loginStep.pause(loginPause))
+    .exec(loginStep.pause(props.loginPause))
     .exec(Dashboard.load(appConfig.baseUrl, defaultHeaders))
 
   val scnCanvas: ScenarioBuilder = scenario(scenarioName("canvas"))
-    .exec(loginStep.pause(loginPause))
+    .exec(loginStep.pause(props.loginPause))
     .exec(Canvas.loadWorkpad(appConfig.baseUrl, defaultHeaders))
 
   setUp(
     scnDiscover
-      .inject(atOnceUsers(400), nothingFor(scnPause))
+      .inject(atOnceUsers(props.maxUsers), nothingFor(props.scnPause))
       .andThen(
         scnDashboard
-          .inject(atOnceUsers(400), nothingFor(scnPause))
-          .andThen(scnCanvas.inject(atOnceUsers(400)))
+          .inject(atOnceUsers(props.maxUsers), nothingFor(props.scnPause))
+          .andThen(scnCanvas.inject(atOnceUsers(props.maxUsers)))
       )
-  ).protocols(httpProtocol).maxDuration(simulationTimeout)
+  ).protocols(httpProtocol).maxDuration(props.simulationTimeout)
 }
